Refactor to move DefaultPublisher into the core
